Replacing `Recommender` to `super()` (#25)
* Replace `Recommender` to `super()`

* Move `_filter_seen` from ALS and Word2Vec to `ANNMixin`
